requirements.txt
gunicorn
dj-database-url
django-filter
whitenoise==3.3.0
psycopg2
Django==2.2.13
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1